color.color -> colors.colorize
