+++ packages/modeling/src/colors/colorize.test.js
@@ -2,13 +2,13 @@ const test = require('ava')
 
 const { geom2, geom3, path2, poly3 } = require('../geometry')
 
-const { color } = require('./index')
+const { colorize } = require('./index')
 
 test('color (rgb on objects)', t => {
   const obj1 = {}
   const obj2 = [{ id: 'a' }, { id: 'b' }]
 
-  const obs = color([1, 0, 0], obj1, obj2)
+  const obs = colorize([1, 0, 0], obj1, obj2)
   const exp1 = { color: [1, 0, 0, 1] }
   const exp2 = { id: 'a', color: [1, 0, 0, 1] }
 
@@ -16,7 +16,7 @@ test('color (rgb on objects)', t => {
   t.deepEqual(obs[0], exp1)
   t.deepEqual(obs[1], exp2)
 
-  const obs3 = color([1, 0, 0], obj1)
+  const obs3 = colorize([1, 0, 0], obj1)
   const exp3 = { color: [1, 0, 0, 1] }
   t.deepEqual(obs3, exp3)
 })
@@ -25,7 +25,7 @@ test('color (rgba on objects)', t => {
   const obj1 = {}
   const obj2 = [{ id: 'a' }, { id: 'b' }]
 
-  const obs = color([1, 1, 0.5, 0.8], obj1, obj2)
+  const obs = colorize([1, 1, 0.5, 0.8], obj1, obj2)
   const exp1 = { color: [1, 1, 0.5, 0.8] }
   const exp2 = { id: 'a', color: [1, 1, 0.5, 0.8] }
 
@@ -40,7 +40,7 @@ test('color (rgba on geometry)', t => {
   const obj2 = path2.fromPoints({closed: false}, [[1, 1]])
   const obj3 = poly3.fromPoints([[0, 0, 0], [1, 0, 0], [1, 1, 0]])
 
-  const obs = color([1, 1, 0.5, 0.8], obj0, obj1, obj2, obj3)
+  const obs = colorize([1, 1, 0.5, 0.8], obj0, obj1, obj2, obj3)
   t.is(obs.length, 4)
   // t.true(geom2.equals(obj0, obs[0]))
   t.not(obj0, obs[0])
