+++ packages/modeling/src/colors/cssColors.js
@@ -1,5 +1,5 @@
 /**
- * @alias module:modeling/color.cssColors
+ * @alias module:modeling/colors.cssColors
  * @see CSS color table from http://www.w3.org/TR/css3-color/
  */
 const cssColors = {
