+++ packages/modeling/src/colors/hslToRgb.js
@@ -8,10 +8,10 @@ const hueToColorComponent = require('./hueToColorComponent')
  * @see http://en.wikipedia.org/wiki/HSL_color_space.
  * @param {...Number|Array} values - HSL or HSLA color values
  * @return {Array} RGB or RGBA color values
- * @alias module:modeling/color.hslToRgb
+ * @alias module:modeling/colors.hslToRgb
  *
  * @example
- * let mysphere = color(hslToRgb([0.9166666666666666, 1, 0.5]), sphere())
+ * let mysphere = colorize(hslToRgb([0.9166666666666666, 1, 0.5]), sphere())
  */
 const hslToRgb = (...values) => {
   values = flatten(values)
